requirements.txt
flask
whoosh
spacy
transformers
torch
python-docx
pymupdf